requirements.txt
paramiko

# ODBC Database connectivity library (required for pyodbc.connect)
pyodbc

# Library to load environment variables from a .env file
python-dotenv
iLibrary
python-dotenv
flet